Feature likes

+++ src/main/java/onboarding/crud/post/controller/PostController.java
@@ -5,14 +5,16 @@ import jakarta.servlet.http.HttpSession;
 import onboarding.crud.post.dto.CreatePostDto;
 import onboarding.crud.post.dto.PostDto;
 import onboarding.crud.post.dto.UpdatePostDto;
-import onboarding.crud.post.entity.Post;
 import onboarding.crud.post.service.PostService;
+import onboarding.crud.user.dto.UserDto;
+import onboarding.crud.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
@@ -22,13 +24,26 @@ public class PostController {
     @Autowired
     private PostService postService;
 
+    @Autowired
+    private UserService userService;
+
     @GetMapping
     public List<PostDto> getAllPosts() {
         return postService.getAllPosts();
     }
 
     @PostMapping
-    public ResponseEntity<?> createPost(@RequestBody CreatePostDto createPostDto) {
+    public ResponseEntity<?> createPost(@RequestBody CreatePostDto createPostDto, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        Object _userId = session.getAttribute("userId");
+        if(_userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
+        }
+        Optional<UserDto> author = userService.getUserById(Long.parseLong(_userId.toString()));
+        if(author.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
+        }
+        createPostDto.setAuthor(author.get().getName());
         try {
             PostDto savedPost = postService.writePost(createPostDto);
             return ResponseEntity.ok(savedPost);
@@ -51,27 +66,52 @@ public class PostController {
     public ResponseEntity<?> likePost(@PathVariable Long id, HttpServletRequest request) {
         HttpSession session = request.getSession();
         Object _userId = session.getAttribute("userId");
-        if (_userId == null) {
+        if(_userId == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
         }
-        String userId = _userId.toString();
-        return ResponseEntity.status(500).body("it will be implemented soon");
+        Optional<UserDto> user = userService.getUserById(Long.parseLong(_userId.toString()));
+        if(user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
+        }
+        if(postService.toggleLikePost(id, user.get().getId())){
+            return ResponseEntity.ok("게시글 좋아요가 성공적으로 변경되었습니다.");
+        }else{
+            return ResponseEntity.status(404).body("해당하는 게시글을 찾을 수 없습니다.");
+        }
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<String> updatePost(@PathVariable Long id, @RequestBody UpdatePostDto updatePostDto) {
-        if (postService.modifyPost(id, updatePostDto)) {
+        if(postService.modifyPost(id, updatePostDto)){
             return ResponseEntity.ok("게시글 수정이 성공적으로 완료되었습니다.");
-        } else {
+        }else{
             return ResponseEntity.status(404).body("게시글을 찾을 수 없습니다.");
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deletePost(@PathVariable Long id) {
-        if (postService.deletePost(id)) {
+    public ResponseEntity<String> deletePost(@PathVariable Long id, HttpServletRequest request) {
+
+        HttpSession session = request.getSession();
+        Object _userId = session.getAttribute("userId");
+        if(_userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
+        }
+        Optional<UserDto> user = userService.getUserById(Long.parseLong(_userId.toString()));
+        if(user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("알맞은 회원으로 로그인이 필요합니다.");
+        }
+        Optional<PostDto> _post = postService.getPostById(id);
+        if(_post.isEmpty()) {
+            return ResponseEntity.status(404).body("게시글을 찾을 수 없습니다.");
+        }
+        PostDto post = _post.get();
+        if(!Objects.equals(post.getAuthor(), user.get().getName())){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("게시글 삭제 권한이 없습니다.");
+        }
+        if(postService.deletePost(id)){
             return ResponseEntity.ok("게시글 삭제가 성공적으로 완료되었습니다.");
-        } else {
+        }else{
             return ResponseEntity.status(404).body("게시글을 찾을 수 없습니다.");
         }
     }
